+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Changed commands:
     - ETA and RP text have been merged into the new "Message" feature:
         - "Message" is entered via Modal and sent to Discord destinations in a separate embed
     - buttons or select menu options for sends, replacing chat letter entry
+        - note **you can SCROLL the select menu** to reveal additional options
     - added icons for send status embeds
     - completely new embed format for Discord
     - added ability to send via webhook
@@ -48,6 +49,7 @@ New commands:
 - `/cco webhook view` - CCO only - shows their webhooks
 - `/cco webhook delete` - CCO only - used to remove a webhook from their list
 - `/cco edit` - CCO only - used to edit an in-progress mission. Presently only the original interaction parameters can be edited (i.e. the fields added via slash command). Editing message and mission type will be implemented in a future update.
+- `/admin_delete_mission` - Admin only - used to manually remove a mission from the database (without cleanup). Intended for unresolvable error situations only.
 - added **training mode** to `cco load` and `cco unload`. This will not affect webhooks but will send discord alerts to the new training channels, and reddit posts to the testing subreddit. Webhooks will be sent as normal, but it's expected trainees will add a training webhook anyway.